move-api-call-for-show(1) Moved api call from Show page to Log component
