+++ src/components/Log.css
@@ -24,3 +24,7 @@
 #log-post {
     height: 50px;
 }
+
+h2 { 
+    text-align: "center"; 
+}
